The scenario editor now genartes goods even if only a preexisting savegame was loaded, and no new game was created during the session.
Some cleanings:
modified trunk/ctp2_code/gs/world/wldgen.cpp

Reduce the memory need for the AI element objects, this should speed up the game a little bit:
modified trunk/ctp2_code/ai/strategy/agents/agent.cpp
modified trunk/ctp2_code/ai/strategy/agents/agent.h
modified trunk/ctp2_code/ai/strategy/goals/Goal.cpp
modified trunk/ctp2_code/ai/strategy/goals/Goal.h
modified trunk/ctp2_code/ai/strategy/scheduler/Plan.cpp
modified trunk/ctp2_code/ai/strategy/scheduler/Plan.h
modified trunk/ctp2_code/ai/strategy/scheduler/scheduler.cpp
modified trunk/ctp2_code/ai/strategy/squads/squad_Strength.h

Fixed some 64 bit porting problems:
modified trunk/ctp2_code/gs/gameobj/CityData.cpp
modified trunk/ctp2_code/gs/gameobj/Happy.cpp
modified trunk/ctp2_code/ui/netshell/ns_accessor.h

Good richness is now saved in the profile.txt so that the Scenario editor regenerates goods when a game was freshly loaded:
modified trunk/ctp2_code/gs/database/profileDB.cpp


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@1070 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/gs/gameobj/Happy.cpp
@@ -938,10 +938,11 @@ double Happy::CalcFeats(Player *p)
 //----------------------------------------------------------------------------
 void Happy::Copy(Happy *copy)
 {
-	memcpy(&m_happiness, &copy->m_happiness, (uint32)&m_crime + sizeof(m_crime) - (uint32)&m_happiness);
+	memcpy(&m_happiness, &copy->m_happiness, (size_t)&m_crime + sizeof(m_crime) - (size_t)&m_happiness);
 	m_timedChanges = copy->m_timedChanges;
 
-	for (size_t i = 0; i < static_cast<size_t>(HAPPY_REASON_MAX); ++i){
+	for (size_t i = 0; i < static_cast<size_t>(HAPPY_REASON_MAX); ++i)
+	{
 		double      data;
 		StringId    name;
 		copy->m_tracker->GetHappiness(static_cast<HAPPY_REASON>(i), data, name);
